refactor: 예약 저장소를 Repository로 분리
