+++ src/main/java/roomescape/ReservationRepository.java
@@ -1,0 +1,7 @@
+package roomescape;
+
+import java.util.List;
+
+public interface ReservationRepository {
+    List<Reservation> findAll();
+}
