+++ src/main/java/roomescape/ReservationService.java
@@ -2,27 +2,18 @@ package roomescape;
 
 import org.springframework.stereotype.Service;
 
-import java.time.LocalDate;
-import java.time.LocalTime;
-import java.util.ArrayList;
 import java.util.List;
 
 @Service
 public class ReservationService {
 
-    private final List<Reservation> reservations = new ArrayList<>();
+    private final ReservationRepository reservationRepository;
 
-    public ReservationService() {
-        initializeReservations();
+    public ReservationService(ReservationRepository reservationRepository) {
+        this.reservationRepository = reservationRepository;
     }
 
     public List<Reservation> findAll() {
-        return List.copyOf(reservations);
-    }
-
-    private void initializeReservations() {
-        reservations.add(new Reservation(1L, "브라운", LocalDate.of(2023, 1, 1), LocalTime.of(10, 0)));
-        reservations.add(new Reservation(2L, "브라운", LocalDate.of(2023, 1, 2), LocalTime.of(11, 0)));
-        reservations.add(new Reservation(3L, "브라운", LocalDate.of(2023, 1, 3), LocalTime.of(12, 0)));
+        return reservationRepository.findAll();
     }
 }
